cluster membership response contains assigned id
switched the response note on cluster membership response to be an ID
for the accpeted client, or an error code if it was denied. The client
side util method returns a new clienthread if the request was
successful, and null if it was not. 
revised serverSelectSubPanel.selectButtonActionPerformed() and
DirectConnectPanel.connectButtonActionPerformed() to reflect the above
change.