Fixed problem where running the game used up all cpu. Also found error.
The problem was that there was an infinite loop with an if statement
inside constantly being evaluated to see if enough time had passed to
update. I fixed this by using Thread.sleep(), although maybe another way
could be used. Also, found a concurrency modification exception thrown
in the reconcile thread in Box. I figured this all out using the eclipse
debugger thread manager which is actually super helpful.